Implement loading at runtime of protocols and skill handlers.
